Move 'show radius' commands implementation to separated file
